refactor: DRY up responses which fetch a single value.
* Fix DictionaryGetField.toString
