+++ integration/dictionary.test.ts
@@ -257,7 +257,7 @@ describe('Integration tests for dictionary operations', () => {
     );
     expect(getResponse).toBeInstanceOf(CacheDictionaryGetField.Hit);
     expect((getResponse as CacheDictionaryGetField.Hit).toString()).toEqual(
-      `${value.substring(0, 32)}...`
+      `Hit: ${value.substring(0, 32)}...`
     );
   });
