+++ src/messages/responses/cache-dictionary-get-field.ts
@@ -1,56 +1,30 @@
 // older versions of node don't have the global util variables https://github.com/nodejs/node/issues/20365
+import * as GetValue from './response-get-value';
 import {TextDecoder} from 'util';
 import {SdkError} from '../../errors/errors';
-import {
-  ResponseBase,
-  ResponseHit,
-  ResponseMiss,
-  ResponseError,
-} from './response-base';
 
 const TEXT_DECODER = new TextDecoder();
 
-export abstract class Response extends ResponseBase {}
+export abstract class Response extends GetValue.Response {}
 
-class _Hit extends Response {
-  private readonly body: Uint8Array;
+export class Hit extends GetValue.Hit {
   private readonly field: Uint8Array;
 
   constructor(body: Uint8Array, field: Uint8Array) {
-    super();
-    this.body = body;
+    super(body);
     this.field = field;
-  }
-  /**
-   * decodes the body into a utf-8 string
-   * @returns string
-   */
-  public valueString(): string {
-    return TEXT_DECODER.decode(this.body);
   }
 
   public fieldString(): string {
     return TEXT_DECODER.decode(this.field);
   }
 
-  public valueUint8Array(): Uint8Array {
-    return this.body;
-  }
-
   public fieldUint8Array(): Uint8Array {
     return this.field;
   }
-
-  public override toString(): string {
-    if (this.valueString().length > 32) {
-      return this.valueString().substring(0, 32) + '...';
-    }
-    return super.toString() + ': ' + this.valueString();
-  }
 }
-export class Hit extends ResponseHit(_Hit) {}
 
-class _Miss extends Response {
+export class Miss extends GetValue.Miss {
   private readonly field: Uint8Array;
 
   constructor(field: Uint8Array) {
@@ -66,13 +40,12 @@ class _Miss extends Response {
     return this.field;
   }
 }
-export class Miss extends ResponseMiss(_Miss) {}
 
-class _Error extends Response {
+export class Error extends GetValue.Error {
   private readonly field: Uint8Array;
 
   constructor(protected _innerException: SdkError, field: Uint8Array) {
-    super();
+    super(_innerException);
     this.field = field;
   }
 
@@ -84,4 +57,3 @@ class _Error extends Response {
     return this.field;
   }
 }
-export class Error extends ResponseError(_Error) {}
